Crate parc

This crate exposes ParentArc<T> which is comparable to an Arc<T> but "strong" references cannot be cloned which allows the ParentArc<T> to lock its weak references and block until all strong references are dropped. Once it is the only reference it can be consummed safely.

This crate is compatible with #![no_std] environnements that provide an allocator.

Structs

ChildArc
Unclonable owned reference to a ParentArc.

LockWeak
Weak reference to a ParentArc.

ParentArc
Owner of multiple atomically reference counted children.

Enums

TryUnwrapError
Errors for TryArcResult

Type Definitions

TryUnwrapResult
Result Type for try_into_inner